Extract artifact panel body
