+++ src/app/features/chat/artifacts/artifact-panel.tsx
@@ -13,11 +13,8 @@ import {
 import {
   type CSSProperties,
   type KeyboardEvent,
-  lazy,
   type PointerEvent,
-  Suspense,
   useCallback,
-  useEffect,
   useLayoutEffect,
   useRef,
   useState,
@@ -28,18 +25,10 @@ import { Tooltip } from '../../../components/common/tooltip'
 import { useDismissibleLayer } from '../../../hooks/useDismissibleLayer'
 import { compactIconButtonClass, popoverPanelClass } from '../../../ui/classes'
 import { cn } from '../../../utils/cn'
-import { HistoricalMarkdownPreview } from './artifact-markdown-preview'
+import { ArtifactPanelBody } from './artifact-panel-body'
 import { formatArtifactSlug } from './artifactFormat'
 import { useArtifactPanelState } from './useArtifactPanelState'
 
-const ArtifactMarkdownEditor = lazy(async () => {
-  const { default: Prism } = await import('prismjs')
-  // @mdxeditor/@lexical code highlighting expects Prism on the browser global.
-  // Install it only on the markdown editor path so HTML/React previews stay isolated.
-  globalThis.Prism = Prism
-  const module = await import('./artifact-markdown-editor')
-  return { default: module.ArtifactMarkdownEditor }
-})
 
 type ArtifactPanelProps = {
   conversationId: string | null
@@ -47,168 +36,6 @@ type ArtifactPanelProps = {
   fullscreen: boolean
   onToggleFullscreen: () => void
   onClose: () => void
-}
-
-function ArtifactPreviewIframe({
-  previewHtml,
-  previewRevision,
-  selectedArtifactKey,
-  setPreviewSource,
-}: {
-  previewHtml: string
-  previewRevision: number
-  selectedArtifactKey: string
-  setPreviewSource: (source: MessageEventSource | null) => void
-}) {
-  const [previewUrl, setPreviewUrl] = useState('')
-
-  useEffect(() => {
-    if (!previewHtml) {
-      setPreviewUrl('')
-      return
-    }
-
-    const nextPreviewUrl = URL.createObjectURL(new Blob([previewHtml], { type: 'text/html' }))
-    setPreviewUrl(nextPreviewUrl)
-    return () => URL.revokeObjectURL(nextPreviewUrl)
-  }, [previewHtml])
-
-  if (!previewUrl) return null
-  return (
-    <iframe
-      ref={(node) => setPreviewSource(node?.contentWindow ?? null)}
-      key={`${selectedArtifactKey}:${previewRevision}:${previewUrl}`}
-      sandbox="allow-scripts allow-forms allow-modals allow-popups"
-      src={previewUrl}
-      className="h-full w-full border-0"
-      title="Artifact preview"
-    />
-  )
-}
-
-function ArtifactPanelBody({
-  fullscreen,
-  panel,
-}: {
-  fullscreen: boolean
-  panel: ReturnType<typeof useArtifactPanelState>
-}) {
-  const {
-    artifacts,
-    artifactLoadError,
-    displayedContent,
-    draft,
-    markdownPreviewEditable,
-    previewError,
-    previewHtml,
-    previewRevision,
-    selectedArtifact,
-    setDraft,
-    setPreviewError,
-    setPreviewSource,
-    setSelectedArtifactId,
-    setView,
-    showingHistoricalVersion,
-    view,
-  } = panel
-  if (artifactLoadError)
-    return (
-      <div className="grid h-full place-items-center px-6 text-center text-[12px] text-[color:var(--danger)]">
-        {artifactLoadError}
-      </div>
-    )
-  if (panel.loadingArtifacts)
-    return (
-      <div className="grid h-full place-items-center px-6 text-center text-[12px] text-[color:var(--muted)]">
-        Loading artifacts…
-      </div>
-    )
-  if (artifacts.length === 0)
-    return (
-      <div className="grid h-full place-items-center px-6 text-center text-[12px] text-[color:var(--muted)]">
-        No artifacts yet.
-      </div>
-    )
-  if (view === 'list') {
-    return (
-      <div className="h-full overflow-y-auto p-2">
-        <div className="grid gap-1">
-          {artifacts.map((artifact) => (
-            <button
-              key={artifact.slug}
-              type="button"
-              className={cn(
-                'rounded-lg px-3 py-2.5 text-left text-[12px] text-[color:var(--muted)] transition-colors hover:bg-[color:var(--surface-hover)] hover:text-[color:var(--text)]',
-                artifact.slug === selectedArtifact?.slug &&
-                  'bg-[color:var(--accent-bg-subtle)] text-[color:var(--text)]',
-              )}
-              onClick={() => {
-                setSelectedArtifactId(artifact.slug)
-                setView('preview')
-              }}
-            >
-              <div className="truncate font-medium">{formatArtifactSlug(artifact.slug)}</div>
-              <div className="mt-0.5 text-[10px] uppercase tracking-[0.12em] text-[color:var(--muted-2)]">
-                {artifact.kind} · v{artifact.version}
-              </div>
-            </button>
-          ))}
-        </div>
-      </div>
-    )
-  }
-  if (view === 'code')
-    return (
-      <textarea
-        className="h-full w-full resize-none overflow-auto bg-[color:var(--panel)] p-3 font-mono text-[12px] leading-5 text-[color:var(--text)] outline-none"
-        value={draft}
-        spellCheck={false}
-        readOnly={showingHistoricalVersion}
-        onChange={(event) => setDraft(event.target.value)}
-      />
-    )
-  if (markdownPreviewEditable) {
-    return (
-      <Suspense
-        fallback={
-          <div className="grid h-full place-items-center text-[12px] text-[color:var(--muted)]">
-            Loading markdown editor…
-          </div>
-        }
-      >
-        <ArtifactMarkdownEditor
-          artifactKey={`${selectedArtifact?.slug}:${selectedArtifact?.version}`}
-          content={draft}
-          diffMarkdown={selectedArtifact?.content ?? ''}
-          fullscreen={fullscreen}
-          onChange={setDraft}
-          onError={setPreviewError}
-        />
-      </Suspense>
-    )
-  }
-  if (view === 'preview' && selectedArtifact?.kind === 'markdown' && showingHistoricalVersion)
-    return <HistoricalMarkdownPreview content={displayedContent} />
-  if (view === 'preview' && selectedArtifact?.kind !== 'markdown') {
-    return (
-      <div className="relative h-full bg-[color:var(--sidebar)]">
-        {previewError ? (
-          <pre className="absolute right-2 bottom-2 left-2 z-10 max-h-32 overflow-auto rounded-lg border border-[color:var(--danger-border)] bg-[color:var(--panel)] p-2 text-[11px] whitespace-pre-wrap text-[color:var(--danger)] shadow-[var(--shadow)]">
-            {previewError}
-          </pre>
-        ) : null}
-        {previewHtml ? (
-          <ArtifactPreviewIframe
-            previewHtml={previewHtml}
-            previewRevision={previewRevision}
-            selectedArtifactKey={`${selectedArtifact?.slug}:${selectedArtifact?.version}:${selectedArtifact?.updatedAt}`}
-            setPreviewSource={setPreviewSource}
-          />
-        ) : null}
-      </div>
-    )
-  }
-  return null
 }
 
 function ArtifactVersionSelect({ panel }: { panel: ReturnType<typeof useArtifactPanelState> }) {
